Rename authentication package to tenant, long overdue
