+++ robozonky-app/src/main/java/com/github/robozonky/app/configuration/OperatingMode.java
@@ -22,8 +22,8 @@ import java.util.function.Consumer;
 
 import com.github.robozonky.api.SessionInfo;
 import com.github.robozonky.api.confirmations.ConfirmationProvider;
-import com.github.robozonky.app.authentication.EventTenant;
-import com.github.robozonky.app.authentication.TenantBuilder;
+import com.github.robozonky.app.tenant.EventTenant;
+import com.github.robozonky.app.tenant.TenantBuilder;
 import com.github.robozonky.app.daemon.DaemonInvestmentMode;
 import com.github.robozonky.app.daemon.operations.Investor;
 import com.github.robozonky.app.events.Events;
