+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/operations/InvestingSession.java
@@ -30,7 +30,7 @@ import com.github.robozonky.api.strategies.InvestmentStrategy;
 import com.github.robozonky.api.strategies.LoanDescriptor;
 import com.github.robozonky.api.strategies.PortfolioOverview;
 import com.github.robozonky.api.strategies.RecommendedLoan;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.common.tenant.Tenant;
 import io.vavr.control.Either;
 import org.slf4j.Logger;
