+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/operations/Purchasing.java
@@ -21,7 +21,7 @@ import java.util.Collection;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.strategies.ParticipationDescriptor;
 import com.github.robozonky.api.strategies.PurchaseStrategy;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 
 public class Purchasing extends StrategyExecutor<ParticipationDescriptor, PurchaseStrategy> {
