+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/operations/PurchasingSession.java
@@ -28,7 +28,7 @@ import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.strategies.ParticipationDescriptor;
 import com.github.robozonky.api.strategies.PurchaseStrategy;
 import com.github.robozonky.api.strategies.RecommendedParticipation;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
