+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/transactions/ParticipationSoldProcessor.java
@@ -21,7 +21,7 @@ import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.TransactionCategory;
 import com.github.robozonky.api.remote.enums.TransactionOrientation;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.app.daemon.LoanCache;
 
 import static com.github.robozonky.app.events.impl.EventFactory.investmentSold;
