+++ robozonky-app/src/main/java/com/github/robozonky/app/delinquencies/DelinquencyNotificationPayload.java
@@ -27,8 +27,8 @@ import java.util.stream.Stream;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.PaymentStatus;
-import com.github.robozonky.app.authentication.EventTenant;
-import com.github.robozonky.app.authentication.TransactionalEventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
+import com.github.robozonky.app.tenant.TransactionalEventTenant;
 import com.github.robozonky.app.daemon.LoanCache;
 import com.github.robozonky.common.jobs.TenantPayload;
 import com.github.robozonky.common.remote.Zonky;
