+++ robozonky-app/src/main/java/com/github/robozonky/app/tenant/Divisor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.app.authentication;
+package com.github.robozonky.app.tenant;
 
 import java.util.concurrent.atomic.LongAdder;
