+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/DaemonInvestmentModeTest.java
@@ -25,7 +25,7 @@ import java.util.concurrent.TimeoutException;
 
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
 import com.github.robozonky.app.ReturnCode;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.app.daemon.operations.Investor;
 import com.github.robozonky.app.runtime.Lifecycle;
 import com.github.robozonky.common.jobs.SimpleJob;
