+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/InvestmentDaemonTest.java
@@ -25,7 +25,7 @@ import com.github.robozonky.api.remote.entities.sanitized.MarketplaceLoan;
 import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.api.strategies.InvestmentStrategy;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.app.daemon.operations.Investor;
 import com.github.robozonky.common.remote.Zonky;
 import org.junit.jupiter.api.Test;
