+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/PurchasingDaemonTest.java
@@ -26,7 +26,7 @@ import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.api.strategies.ParticipationDescriptor;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.app.daemon.transactions.SoldParticipationCache;
 import com.github.robozonky.common.remote.Zonky;
 import org.junit.jupiter.api.Test;
