+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/operations/PurchasingSessionTest.java
@@ -31,7 +31,7 @@ import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.api.strategies.ParticipationDescriptor;
 import com.github.robozonky.api.strategies.PurchaseStrategy;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.EventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.common.tenant.RemotePortfolio;
 import org.assertj.core.api.Condition;
