+++ robozonky-app/src/test/java/com/github/robozonky/app/delinquencies/CategoryTest.java
@@ -25,8 +25,8 @@ import com.github.robozonky.api.notifications.LoanNowDelinquentEvent;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.EventTenant;
-import com.github.robozonky.app.authentication.TransactionalEventTenant;
+import com.github.robozonky.app.tenant.EventTenant;
+import com.github.robozonky.app.tenant.TransactionalEventTenant;
 import com.github.robozonky.common.remote.Zonky;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
