+++ robozonky-app/src/test/java/com/github/robozonky/app/tenant/ZonkyApiTokenSupplierTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.app.authentication;
+package com.github.robozonky.app.tenant;
 
 import java.time.Duration;
 import java.time.Instant;
